extend via value class implicits, i.e. no runtime overhead
